requirements.txt
# FastAPI and web framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
alembic>=1.12.0

# AI and ML
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0

# Background processing
celery>=5.3.0
redis>=5.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
aiofiles>=23.0.0

# Development
pytest>=7.0.0